requirements.txt
fastapi
uvicorn
websockets
python-jose
passlib
python-multipart
pydantic
sqlalchemy
alembic
bcrypt
python-dotenv
pyjwt